Remove FileTypeSelection
- Disable options appropriately
