+++ package.json
@@ -108,7 +108,7 @@
     "dependencies": {
         "@babel/polyfill": "^7.2.5",
         "@timilsinabishal/react-components": "https://github.com/the-deep/react-components",
-        "@togglecorp/fujs": "^1.0.1",
+        "@togglecorp/fujs": "^1.1.0",
         "@togglecorp/ravl": "^1.2.0",
         "@togglecorp/react-rest-request": "^1.3.1",
         "body-parser": "^1.18.3",
