+++ src/redux/initial-state/dev-lang.json
@@ -1639,7 +1639,7 @@
             "noFiltersMessage": 1139
         },
         "addLeads.tabular": {
-            "title": 15007,
+            "title": -2259039,
             "delimiterLabel": 15004,
             "fileTypeLabel": 15005,
             "skipLabel": 15009,
