--- src/views/LeadAdd/LeadFormItem/LeadTabular/FileTypeSelectionPage/styles.scss
@@ -1,8 +1,0 @@
-@import '~base-scss/base';
-
-.form {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-    overflow-y: auto;
-}
